Drop RateIt class. Move logic to the controller instead.
